shared_ptr's as far as the eye can see
Also (believe) I have ranking working and I have a function that checks if the last played card was a card that has a rule we're implementing;

Black Queen forces 5
Two forces 2
8 Forces wait
Jack changes direction

The last 2 don't actually do anything yet but they're there.

NEED a game manager to tell us whose turn it is, and when the last card was played. Otherwise, black queens and twos are gonna crash the game cause everyone will pick up and we'll run out of cards.