SFML!
Adding SFML so that I can give the game some visuals rather than just being console based.

In Advanced Games Engineering (SET10110), Sam gave us a starter project for week 1 that used CMake to download a version of SFML and set up all of the dependencies etc. However, this no longer works for some reason, and as I don't really understand his CMake file, I've just downloaded the latest SFML version and put it into the project, and altered the CMakeLists file to set the dependencies for that.