I believe I now have a method that sorts out all of the textures (face up/down) and positions of cards with a single call.
At the moment, I'm passing it the hand of the current player as a parameter though. As the game progresses this'll probably need to be changed so that it's being given only the hand of the actual human user. Of course, playing this with more than one human will need rotation in some way, but that can be figured out later; It's not a priority