EPILEPSY WARNING :warning:
The window colour flickers between green and black (an error somewhere with how often the screen is being displayed or something), but I've got clicking on sprites working.

I just have something in the while window is open loop that loops through the cards in the player's hand and if you click on a card, it becomes face down.

That's pointless, but it was just something to do to test the clicking on a sprite was working. It means I can use that code to select which card to play with the mouse when that's ready to happen