Fixed texture problem
Turns out that SetTextures() wasn't actually turning everything back to being face up.

So, I moved its contents back up to SetUpDeck() so that it's all done in a single call, and deleted SetTextures().

Each card object now has a texture variable, which is set in SetUpDeck(). So, when UpdatePositionsAndTextures() is called, it now can just tell each card to revert to the texture that it has stored.