Epilepsy issue resolved
You can now click on the cards in your hand, and if they're playable (in relation to their suit/type), it SHOULD add it to the last card array, and draw it in the middle of the screen.

The previous last card is moved to the stack, and the card you just clicked on should be removed from your hand.

However, this is weirdly only working if it is the last card in your hand! I have no idea why this is at the moment.