Seem to have fixed the face down issue and turned the ranking system into an object to keep it tidy
However, I'm now getting a couple of other problems. Playing cards seems to (sometimes, not sure if it's all the time) swap the played card with the card to be played on top of. The card being played on top of is jumping into one of the players hands. There's also an issue with not being allowed to pick up sometimes, probably a conflict between ConsultRules() and can_play_checker() setting the canPlay and canPickUp bools.