I think I fixed it!! :smiley:
Next issue:

Checking who played the last card (used for special cards) causes the rule to be enforced multiple times if the card remains as the last played card as we iterate through the list of players.

Need a way to set it to be inactive after the punishment has been dealt.