Games are now fully automatic :+1:
There was a bug where the mouse had to be active within the window for the game to continue before, but now it works and can run in the background.

Furthermore, GameRecords.txt now tracks game numbers so I can see where records are from rematches. Doesn't really matter now but could come in handy later