Taking Jyoti's Formatting Advice
At the end of her lecture on Monday, she decided to give us a few tips about writing our dissertation. One of those was to ensure that anything before the introduction was page-numbered in Roman numerals, and the introduction section starts on the official page 1.

I've emailed her to see what she recommends for post-conclusion pages.